--- a/packages/superset-ui-plugins-demo/.storybook/webpack.config.js
+++ b/packages/superset-ui-plugins-demo/.storybook/webpack.config.js
@@ -1,7 +1,4 @@
 const path = require('path');
-const SpeedMeasurePlugin = require('speed-measure-webpack-plugin');
-
-const smp = new SpeedMeasurePlugin();
 
 const BABEL_TYPESCRIPT_OPTIONS = {
   presets: [
@@ -24,7 +21,8 @@ module.exports = async ({ config }) => {
   config.resolve = config.resolve || {};
   config.resolve.extensions = ['.tsx', '.ts', '.jsx', '.js'];
 
-  config.module.noParse = /core[-]js|jquery|lodash|moment|mathjs|d3([-].+)?/;
+  // Avoid parsing large libraries to speed up build
+  config.module.noParse = /core[-]js|jquery|lodash|moment|mathjs|d3([-][a-z]+)?/;
 
   // To enable live debugging of other packages when referring to `src`
   config.module.rules.push({
@@ -60,5 +58,5 @@ module.exports = async ({ config }) => {
     },
   };
 
-  return smp.wrap(config);
+  return config;
 };