--- a/packages/superset-ui-plugins-demo/.storybook/webpack.config.js
+++ b/packages/superset-ui-plugins-demo/.storybook/webpack.config.js
@@ -61,8 +61,10 @@ module.exports = async ({ config }) => {
     },
   };
 
-  console.log('process.env.RUNNING_CONTEXT', process.env.RUNNING_CONTEXT);
-  config.parellism = process.env.RUNNING_CONTEXT === 'netlify' ? 1 : 100;
+  if (process.env.RUNNING_CONTEXT === 'netlify') {
+    config.paralellism = 1;
+    config.devtool = false;
+  }
 
   return config;
 };