--- a/packages/superset-ui-plugins-demo/.storybook/webpack.config.js
+++ b/packages/superset-ui-plugins-demo/.storybook/webpack.config.js
@@ -22,12 +22,12 @@ const SIBLING_PACKAGES_PATH_REGEXP = new RegExp(
 module.exports = async ({ config }) => {
   config.resolve = config.resolve || {};
   config.resolve.extensions = ['.tsx', '.ts', '.jsx', '.js'];
-  config.resolve.alias = {
-    ...config.resolve.alias,
-    d3$: path.resolve(__dirname, '../../../node_modules/d3/d3.min.js'),
-    nvd3$: path.resolve(__dirname, '../../../node_modules/nvd3/build/nv.d3.min.js'),
-    'datatables.net$': path.resolve(__dirname, '../../../node_modules/datatables.net/js/jquery.dataTables.min.js'),
-  }
+  // config.resolve.alias = {
+  //   ...config.resolve.alias,
+  //   d3$: path.resolve(__dirname, '../../../node_modules/d3/d3.min.js'),
+  //   nvd3$: path.resolve(__dirname, '../../../node_modules/nvd3/build/nv.d3.min.js'),
+  //   'datatables.net$': path.resolve(__dirname, '../../../node_modules/datatables.net/js/jquery.dataTables.min.js'),
+  // }
 
   config.plugins.push(new webpack.IgnorePlugin(/^\.\/locale$/, /moment$/));
   // Avoid parsing large libraries to speed up build
@@ -61,14 +61,9 @@ module.exports = async ({ config }) => {
     }],
   });
 
-  // config.optimization = {
-  //   splitChunks: {
-  //     chunks: 'async'
-  //   },
-  // };
-
   if (process.env.RUNNING_CONTEXT === 'netlify') {
     config.devtool = false;
+    config.cache = false;
   }
 
   return config;