--- a/packages/superset-ui-plugins-demo/.storybook/webpack.config.js
+++ b/packages/superset-ui-plugins-demo/.storybook/webpack.config.js
@@ -1,5 +1,9 @@
 const path = require('path');
+const TerserPlugin = require('terser-webpack-plugin');
 const webpack = require('webpack');
+
+const SpeedMeasurePlugin = require("speed-measure-webpack-plugin");
+const smp = new SpeedMeasurePlugin();
 
 const BABEL_TYPESCRIPT_OPTIONS = {
   presets: [
@@ -29,7 +33,7 @@ module.exports = async ({ config }) => {
     'datatables.net$': path.resolve(__dirname, '../../../node_modules/datatables.net/js/jquery.dataTables.min.js'),
   }
 
-  // config.plugins.push(new webpack.IgnorePlugin(/^\.\/locale$/, /moment$/));
+  config.plugins.push(new webpack.IgnorePlugin(/^\.\/locale$/, /moment$/));
   // Avoid parsing large libraries to speed up build
   config.module.noParse = /jquery|moment/;
 
@@ -53,6 +57,7 @@ module.exports = async ({ config }) => {
   });
 
   config.module.rules.push({
+    include: path.resolve(__dirname, '../storybook'),
     exclude: /node_modules/,
     test: /\.tsx?$/,
     use: [{
@@ -60,6 +65,17 @@ module.exports = async ({ config }) => {
       options: BABEL_TYPESCRIPT_OPTIONS,
     }],
   });
+
+  config.optimization = config.optimization || {};
+  config.optimization.splitChunks = {
+    chunks: 'async'
+  };
+  config.optimization.minimizer = [
+    new TerserPlugin({
+      parallel: true,
+      extractComments: true,
+    }),
+  ];
 
   if (process.env.RUNNING_CONTEXT === 'netlify') {
     config.devtool = false;