--- a/packages/superset-ui-plugins-demo/package.json
+++ b/packages/superset-ui-plugins-demo/package.json
@@ -9,6 +9,7 @@
     "demo:build": "build-storybook -o _gh-pages",
     "demo:publish": "gh-pages -d _gh-pages",
     "deploy-demo": "npm run demo:clean && npm run demo:build && npm run demo:publish && npm run demo:clean",
+    "build:dll": "webpack --config library.webpack.config.js",
     "storybook": "start-storybook -p 9001"
   },
   "repository": {
@@ -46,7 +47,8 @@
     "@babel/core": "^7.4.3",
     "@babel/plugin-syntax-dynamic-import": "^7.2.0",
     "babel-loader": "^8.0.5",
-    "gh-pages": "^2.0.1"
+    "gh-pages": "^2.0.1",
+    "terser-webpack-plugin": "1.3.0"
   },
   "peerDependencies": {
     "@superset-ui/chart": "^0.11.13",