--- a/packages/superset-ui-plugins-demo/storybook/stories/legacy-plugin-chart-country-map/index.js
+++ b/packages/superset-ui-plugins-demo/storybook/stories/legacy-plugin-chart-country-map/index.js
@@ -1,8 +1,8 @@
-import CountryMapChartPlugin from '../../../../superset-ui-legacy-plugin-chart-country-map';
-import Stories from './Stories';
+// import CountryMapChartPlugin from '../../../../superset-ui-legacy-plugin-chart-country-map';
+// import Stories from './Stories';
 
-new CountryMapChartPlugin().configure({ key: 'country-map' }).register();
+// new CountryMapChartPlugin().configure({ key: 'country-map' }).register();
 
-export default {
-  examples: [...Stories],
-};
+// export default {
+//   examples: [...Stories],
+// };