--- a/packages/superset-ui-plugins-demo/storybook/stories/legacy-plugin-chart-country-map/index.js
+++ b/packages/superset-ui-plugins-demo/storybook/stories/legacy-plugin-chart-country-map/index.js
@@ -1,3 +1,5 @@
+// Disable temporarily
+
 // import CountryMapChartPlugin from '../../../../superset-ui-legacy-plugin-chart-country-map';
 // import Stories from './Stories';