--- a/packages/superset-ui-plugins-demo/storybook/stories/legacy-plugin-chart-map-box/index.js
+++ b/packages/superset-ui-plugins-demo/storybook/stories/legacy-plugin-chart-map-box/index.js
@@ -1,3 +1,5 @@
+// Disable temporarily
+
 // import MapBoxChartPlugin from '../../../../superset-ui-legacy-plugin-chart-map-box';
 // import Stories from './Stories';