--- a/packages/superset-ui-plugins-demo/storybook/stories/preset-chart-xy/BoxPlot/index.js
+++ b/packages/superset-ui-plugins-demo/storybook/stories/preset-chart-xy/BoxPlot/index.js
@@ -1,5 +1,5 @@
-import { BoxPlotChartPlugin as LegacyBoxPlotChartPlugin } from '../../../../../superset-ui-preset-chart-xy/src/legacy';
-import { BoxPlotChartPlugin } from '../../../../../superset-ui-preset-chart-xy/src';
+import { BoxPlotChartPlugin as LegacyBoxPlotChartPlugin } from '../../../../../superset-ui-preset-chart-xy/esm/legacy';
+import { BoxPlotChartPlugin } from '../../../../../superset-ui-preset-chart-xy';
 import Stories from './stories/Basic';
 import LegacyStories from './stories/Legacy';
 import { BOX_PLOT_PLUGIN_LEGACY_TYPE, BOX_PLOT_PLUGIN_TYPE } from './constants';