🏠 Internal

* Avoid parsing large libraries and reduce optimization aggressiveness to trade for speed and hopefully reduce netlify timeout.

Stats before
```
 SMP  ⏱
General output time took 1 min, 51.81 secs

 SMP  ⏱  Plugins
CaseSensitivePathsPlugin took 47.77 secs
TerserPlugin took 47.23 secs
NormalModuleReplacementPlugin took 16.57 secs
ProgressPlugin took 3.91 secs
HtmlWebpackPlugin took 0.068 secs
DefinePlugin took 0 secs

 SMP  ⏱  Loaders
modules with no loaders took 1 min, 2.87 secs
  module count = 2488
file-loader took 38.43 secs
  module count = 86
babel-loader took 25.066 secs
  module count = 184
css-loader, and
postcss-loader took 17.6 secs
  module count = 30
style-loader, and
css-loader, and
postcss-loader took 0.113 secs
  module count = 30
```

Stats after
```
   SMP  ⏱
General output time took 48.28 secs

 SMP  ⏱  Plugins
CaseSensitivePathsPlugin took 11.058 secs
NormalModuleReplacementPlugin took 1.9 secs
ProgressPlugin took 0.711 secs
HtmlWebpackPlugin took 0.069 secs
DefinePlugin took 0.001 secs

 SMP  ⏱  Loaders
modules with no loaders took 14.52 secs
  module count = 909
file-loader took 12.98 secs
  module count = 74
babel-loader took 8.98 secs
  module count = 161
css-loader, and
postcss-loader took 4.56 secs
  module count = 25
style-loader, and
css-loader, and
postcss-loader took 0.052 secs
  module count = 28
```